// src/notification/notification.enum.ts
export enum NotificationType {
  REGISTER_ACTIVITY = 'REGISTER_ACTIVITY',
  NOTIFY_BEFORE_EVENT_START = 'NOTIFY_BEFORE_EVENT_START',
  CHECKIN_COMPLETE = 'CHECKIN_COMPLETE',
  CHECKOUT_COMPLETE = 'CHECKOUT_COMPLETE',
  CANCEL_COMPLETE = 'CANCEL_COMPLETE',
  NOTIFY_BEFORE_CHECKOUT = 'NOTIFY_BEFORE_CHECKOUT',
  REDEEM_COMPLETE = 'REDEEM_COMPLETE',
  EVOUCHER_EXP = 'EVOUCHER_EXP',
  DONATION_SUCCESS = 'DONATION_SUCCESS',
  RECEIVE_COIN = 'RECEIVE_COIN',
  APPROVE_EVENT = 'APPROVE_EVENT',
  BROADCAST_EVENT = 'BROADCAST_EVENT',
  MY_EVENT = 'MY_EVENT',
}

export enum NotificationBodyLocKey {
  MY_EVENT = 'MY_EVENT',
  REDEEM_HISTORY = 'REDEEM_HISTORY',
  DONATION_HISTORY = 'DONATION_HISTORY',
  TRANSACTION_HISTORY = 'TRANSACTION_HISTORY',
}